Disable start button
While the clock was running a user was able to click on the start
button a second time which caused the setInterval function to run
again.

To resolve the problem a class was added to the start button in
the HTML so that Javascript could target the element and to
disable the button.

A variable was added called running to know when the timer was
running or not.