Improved positioning
Padding was removed from the article so that flexbox could be used
to center the elements and leave a small border around the edge to
display the background image.

I changed the height and width of main elements so that they would
be a bit smaller than the height and width of the article element.